requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
Flask-PyMongo==2.3.0
pymongo==4.7.2
python-dotenv==1.0.0
requests==2.31.0
